Manual test datasets-datamanagers integration (partial) (#1469)
Manual test datasets-datamanagers integration (partial): take datasets,
datamanagers, try run some notebooks (and variations of notebooks for
shaking the datasets/datamanagers).

Add fixes in library (*not* intended to adapt the notebooks)

No issue number associated

Current PR adds some fixes coming from WIP tests of medical folder
dataset's notebook:
- adapt node's `Round` class to prepare `controller_kwargs` as expected
by new `Controller` classes in the `Dataset.complete_initialization`
- adds more explicit messages in torch datamanager's dataset wrapper and
scikit-learn's dataloader batch iterator for bad types of data in the
sample. This is raised in the case it is not caught by the dataset (et
transform is not a dict, just a callable, for MedicalFolderDataset)
- fix default transforms for torch training plan: need to convert to
float (float32) to avoid mismatch error: in the case sample (input) data
is float64, converter would keep it while pytorch model uses by default
float32
- fix DLP usage in MedicalFolderController

For info, manual tests done for this PR include:
- 101 notebook
- `tutorials/medical/medical-image-segmentation-unet-library.ipynb`
testing transforms as callable and as dict of callables.
-
`tutorials/scikit-learn/01_sklearn_MNIST_classification_tutorial.ipynb`
with MNIST and with IXI/MedicalFolderDataset. Tested most error cases on
data format (multiple modalities, bad modality data type, bad modality
data ndim)
- `tutorials/medical/medical-image-segmentation-unet-library.ipynb`
using a DLP which changes all modality names (with adapted notebook +
manual craft of dataset/DLP as CLI/GUI are not ready for it)


---------

+++ fedbiomed/common/dataset/_native_dataset.py
@@ -26,7 +26,7 @@ class NativeDataset(Dataset):
         DataReturnFormat.TORCH: lambda x: (
             T.ToTensor()(x)  # In case the input is a PIL image or ndarray image
             if isinstance(x, (Image.Image, np.ndarray))
-            else torch.tensor(x)
+            else torch.tensor(x).float()
         ),
     }
